% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{add_temperature}
\alias{add_temperature}
\title{convert time to temperature}
\usage{
add_temperature(signal_data, time_range)
}
\arguments{
\item{signal_data}{tibble with variable t and pyro_temp (measured temperature)}

\item{interpolation_times}{times at for which the temperatures are required}
}
\value{
tibble: signal_data with additional variable \emph{smoothed_temp}
}
\description{
convert time to temperature
}
