% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{calc_method_trace}
\alias{calc_method_trace}
\title{adds a trace to st-result-plot for a calc_method}
\usage{
calc_method_trace(
  p,
  cm,
  trace,
  estimates = NULL,
  p_size = 6,
  sl = TRUE,
  fit_lm = FALSE,
  farbe = "black"
)
}
\arguments{
\item{p}{plot_ly object}

\item{cm}{string specifying the calc_method}

\item{trace}{name/legend for the trace}

\item{estimates}{data-frame}

\item{p_size}{point-size}

\item{sl}{logical: showlegend}

\item{fit_lm}{logical: add lm-model-fit to trace}

\item{farbe}{color of the trace}
}
\value{
plot_ly-object
}
\description{
adds a trace to st-result-plot for a calc_method
}
