% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{convert_to_temp}
\alias{convert_to_temp}
\title{convert time to temp}
\usage{
convert_to_temp(time, tevi_data)
}
\arguments{
\item{time}{vector: time-points to convert to temp}

\item{tevi_data}{tibble with vars \emph{t} and \emph{smoothed_temp}}
}
\value{
vector with temperatur
}
\description{
this function takes a time-vector and converts it to a temperature-vector;
it uses the smoothed temperature-values from raw_tevi_data:
if the time is outside the experimental-time-range (specified in import-tevi-data)
it returns NA-values
}
