% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{estimate_signal_spectrum}
\alias{estimate_signal_spectrum}
\title{Estimate Spectrum from Signal}
\usage{
estimate_signal_spectrum(
  signal_data,
  signal_name,
  frame_rate,
  spans = c(3, 3),
  taper = 0.1
)
}
\arguments{
\item{signal_data}{tibble with signal}

\item{signal_name}{char: name of the signal-variable}

\item{frame_rate}{sample-rate in Hz}

\item{spans}{vector of odd integers giving the widths of modified Daniell smoothers to be used to smooth the periodogram.}

\item{taper}{specifies the proportion of data to taper}
}
\value{
tibble with vars: f, spec and fc_amp
}
\description{
Estimate Spectrum from Signal
}
