% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{fftc}
\alias{fftc}
\title{calculate fourier-koeffs. with fft}
\usage{
fftc(data, signal, sr)
}
\arguments{
\item{data}{tibble with var "signal"}

\item{signal}{character: name of signal}

\item{sr}{samplerate}
}
\value{
tibble with vars f (Kreisfrequenz, d.h. cycles per unit-time), fc, fc_amp and fc_arg and spec = N * fc_amp^2
}
\description{
The function \code{fft} returns the \bold{unnormalized} Fourier-Coefficients
for the Fourier-Frequencies 0/N, ... (N-1)/N, where N = T/dT ist the Number of samples.
}
