% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{fit_lorentz}
\alias{fit_lorentz}
\title{fit_lorentz}
\usage{
fit_lorentz(data, signal, sr = 400)
}
\arguments{
\item{data}{tibble with columns f and fc_amp}

\item{signal}{character}

\item{sr}{sample-rate}
}
\value{
list(fit_params, fitted[[f, lf_amp]]) NULL if nls did not converge
}
\description{
fit_lorentz
}
