% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{gen_example_data}
\alias{gen_example_data}
\title{Simuliere Signal}
\usage{
gen_example_data(
  T = 10,
  sr = 400,
  dc = 0.1,
  f0 = 30,
  fs = 0,
  noise_sd = 0.1,
  wr = c(0, 10),
  bp = c(0, 400)
)
}
\arguments{
\item{sr}{sample-rate}

\item{dc}{damping-constant}

\item{f0}{signal-freq}

\item{fs}{freq-shift}

\item{noise_sd}{sd of additive nois}

\item{wr}{window-range}

\item{bp}{band-pass-range}
}
\value{
tibble with vars: t, sig, f, fc, fc_amp, fc_phase, sig_re (reconstructed sig after bp), equal
}
\description{
Simuliere Signal
}
