% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{get_brush_range}
\alias{get_brush_range}
\title{convert brush from plotOutput to numeric vector}
\usage{
get_brush_range(brush, failure_massage = "select range by brushing plot")
}
\arguments{
\item{brush}{input-object}

\item{failure_massage}{string: message to display in case there is no brush}
}
\value{
numeric vector with the range of the brushed area
}
\description{
convert brush from plotOutput to numeric vector
}
