% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{get_fc}
\alias{get_fc}
\title{calculate fourier-koeffs. with fft}
\usage{
get_fc(data, signal, sr)
}
\arguments{
\item{data}{tibble with var "signal"}

\item{signal}{character: name of signal}

\item{sr}{samplerate}
}
\value{
tibble with vars f, fc, fc_amp and fc_arg
}
\description{
The function \code{fft} returns the \bold{unnormalized} Fourier-Coefficients!
To be consistent with Schlittgen and
}
