% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{import_tevi_data}
\alias{import_tevi_data}
\title{loads tevi-data}
\usage{
import_tevi_data(file_path)
}
\arguments{
\item{file_path}{path as supplied from fileInput}
}
\value{
a tibble with all signals + vars: \emph{t, smoothed_temperature}
}
\description{
loads tevi-data
}
