% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{lorentz_amps}
\alias{lorentz_amps}
\title{calculate lorentz-amplitudes from fitted model}
\usage{
lorentz_amps(freqs, lf_model)
}
\arguments{
\item{freqs}{a tibble with a variable f}

\item{lf_model}{a nls model returned from fit_lorentz}
}
\value{
tibble with two variables: f and lf_amp
}
\description{
calculate lorentz-amplitudes from fitted model
}
