% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{lorentz_parameters}
\alias{lorentz_parameters}
\title{Extract the fitted parameters from fitted lorentz-model}
\usage{
lorentz_parameters(lf_model)
}
\arguments{
\item{lf_model}{a nls-object as returned from fit_lorentz}
}
\value{
named vector
}
\description{
Extract the fitted parameters from fitted lorentz-model
}
