% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearest_node_grid}
\alias{nearest_node_grid}
\title{Index of the node nearest to each grid cell}
\usage{
nearest_node_grid(coordNodes, resolutionValue, zoomValue, increase)
}
\arguments{
\item{coordNodes}{n x 2 matrix with the normalised coordinates of the nodes.}

\item{resolutionValue}{side of the square grid.}

\item{zoomValue}{coordinate of the lower-left corner of the grid.}

\item{increase}{grid step.}
}
\value{
Integer matrix resolutionValue x resolutionValue with the
  (1-based) index of the node nearest to each cell.
}
\description{
Used by the GUI to report which gene a region selected on the landscape
belongs to. Brute force in O(res^2 * n), exact and cheap: for a network of
1200 nodes on a 240 x 240 grid this is about 70 million comparisons.
}
\keyword{internal}
