% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{regular_ts}
\alias{regular_ts}
\title{checks if sampling-interval is equidistant and no
observations are missing}
\usage{
regular_ts(data, signal, sr)
}
\arguments{
\item{data}{tibble with var t}

\item{signal}{character: Name of the signal to check}

\item{sr}{samplerate}
}
\value{
boolean: true if ts is regular
}
\description{
checks if sampling-interval is equidistant and no
observations are missing
}
