% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{to_surface_tension}
\alias{to_surface_tension}
\title{convert frequency to surface tension}
\usage{
to_surface_tension(freq, mass)
}
\arguments{
\item{freq}{dominant frequency of oscillating drop in Hz}

\item{mass}{Mass of drop in g}
}
\value{
surface tension of liquid alloy in  Nm
}
\description{
convert frequency to surface tension
}
