% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{to_viscosity}
\alias{to_viscosity}
\title{convert damping-constant to alloy-viscosity}
\usage{
to_viscosity(d, m, r)
}
\arguments{
\item{d}{damping-constant in Hz}

\item{m}{Mass in g}

\item{r}{Radius in mm}
}
\value{
alloy-viscosity in mPa * s
}
\description{
convert damping-constant to alloy-viscosity
}
