% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{bp_filter}
\alias{bp_filter}
\title{Band-Pass-Filter signal}
\usage{
bp_filter(sig_data, signal, bp, sr)
}
\arguments{
\item{sig_data}{tibble with var signal}

\item{signal}{character: name of signal to be filtered}

\item{bp}{numeric vector: lower and upper bp-Freqs}

\item{sr}{samplerate}
}
\value{
tibble with
}
\description{
Band-Pass-Filter signal
}
