% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{bp_filter}
\alias{bp_filter}
\title{Band-Pass-Filter signal}
\usage{
bp_filter(sig_data, signal_name, bp, sr)
}
\arguments{
\item{sig_data}{tibble with var signal}

\item{signal_name}{var-name of the signal}

\item{bp}{numeric vector: lower and upper bp-Freqs}

\item{sr}{samplerate}
}
\value{
tibble with vars t and signal_name
}
\description{
The function calculates the Fourier-Coefficients with fft,
sets all Fourier-Coefficients not within the bp-range to 0
and inverses the fft.
}
