% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{fit_lorentz}
\alias{fit_lorentz}
\title{fit_lorentz}
\usage{
fit_lorentz(fc_data, c0)
}
\arguments{
\item{fc_data}{tibble with columns f and fc_amp, wobei f Werte bis max. der Nyquist-Freq enthält}
}
\value{
list(fit_params, fitted[[f, lf_amp]]) NULL if nls did not converge
}
\description{
fit_lorentz
}
