% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{fit_lorentz}
\alias{fit_lorentz}
\title{fit_lorentz}
\usage{
fit_lorentz(fc, c0)
}
\arguments{
\item{fc}{tibble with columns f and fc_amp, wobei f Werte bis max. der Nyquist-Freq enthält}

\item{c0}{numeric vector c(A, f0, d) with start-values for nls}
}
\value{
list(fit_params, fitted[[f, lf_amp]]) NULL if nls did not converge
}
\description{
fit_lorentz
}
