% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{fit_lorentz}
\alias{fit_lorentz}
\title{fit_lorentz}
\usage{
fit_lorentz(data, sr = 400, c0 = c(A = 1000, f0 = 30, g = 0.1))
}
\arguments{
\item{data}{tibble with columns f and fc_amp}

\item{sr}{sample-rate}

\item{c0}{vector with start-parameters}
}
\value{
list(fit_params, fitted[[lf_amp]]) NULL if nls did not converge
}
\description{
fit_lorentz
}
