% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{gen_spec_plot}
\alias{gen_spec_plot}
\title{generate spectrogram-plot}
\usage{
gen_spec_plot(
  est_spec,
  lfit,
  scale = "log10",
  bp,
  type_choosen = "fft",
  sample_rate,
  plot_type = "ggplot"
)
}
\arguments{
\item{est_spec}{tibble: estimated spectrum}

\item{lfit}{lfit-model}

\item{scale}{string: raw/log10}

\item{bp}{numeric vector: bandpass-limits}

\item{type_choosen}{fft/spectrum}

\item{sample_rate}{sample-rate}

\item{plot_type}{ggplot/plot_ly}
}
\value{
plot-object
}
\description{
generate spectrogram-plot
}
