% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gen_spec_plot}
\alias{gen_spec_plot}
\title{generate spectrogram-plot}
\usage{
gen_spec_plot(est_spec, lfit_models, scale = "log10", sample_rate)
}
\arguments{
\item{est_spec}{tibble: estimated spectrum}

\item{lfit_models}{lfit-models}

\item{scale}{string: raw/log10}

\item{sample_rate}{sample-rate}
}
\value{
plot-object
}
\description{
generate spectrogram-plot
}
