% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gen_spec_plot}
\alias{gen_spec_plot}
\title{generate spectrogram-plot}
\usage{
gen_spec_plot(est_spec, lfit_model)
}
\arguments{
\item{est_spec}{tibble: estimated spectrum}

\item{lfit_model}{return model form lorentz-fit}
}
\value{
plot-object
}
\description{
generate spectrogram-plot
}
