% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{get_dom_freq}
\alias{get_dom_freq}
\title{get dominant freq and corresponding amplitude in signal}
\usage{
get_dom_freq(fc_data)
}
\arguments{
\item{fc_data}{datensatz mit variablen f und fc_amp}
}
\value{
tibble with variables f and fc_amp
(fc_amp is the signal-amplitude of the dominant frequency!
Not the amplitude of the Fourie-Coefficient!)
}
\description{
Ermittelt die dominante Frequenz (die Frequenz mit maximaler Amplitude)
eines Signals aus dem Periodogramm.
}
\details{
Nur Frequenzen ab 0.1 Hz kommen in Frage!

Ist das signal ein reines cosinoid, d.h. eine evtl. verrauschte
harmonische Schwingung so liefert die Theorie: Der Betrag des betragsmäßig
größten (normierten) Fourierkoeffizienten ist die halbe Amplitude des Signals.
Aus diesem Grund wird der zurückgegebene Wert der Amplitude mit 2
multipliziert.
}
