% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{get_spectrum}
\alias{get_spectrum}
\title{get_spectrum}
\usage{
get_spectrum(data, signal, sr)
}
\arguments{
\item{data}{tibble with variable sig}

\item{sr}{sample-rate}
}
\value{

}
\description{
get_spectrum
}
