% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEVIui.R
\name{LEVIui}
\alias{LEVIui}
\title{LEVIui}
\usage{
LEVIui(browser)
}
\arguments{
\item{browser}{This argument is necessary to launch Levi GUI. To launch Levi
in the web browser the argument required "TRUE". To launch Levi in the R
environment the argument required "FALSE". The default is "FALSE"}
}
\value{
return a GUI
}
\description{
Launch the Levi Graphical User Interface (GUI) in local machine.
}
\details{
This function launch the LEVI Graphical User Interface. The
interface provides the same tools available in the script mode. There are
two tools only available in the user interface: 1) Selection of area from
heatmap to calculate the gene expression levels in the area selected;
2) Selection of the genes in some specific area from the image.
}
\examples{
LEVIui(browser)
#LEVIui(browser=TRUE)  #Launch Levi to Browser.
#LEVIui(borwser=FALSE) #Launch Levi to R environment.

}
\author{
José Rafael Pilan <rafael.pilan@unesp.br> &
Isabelle Mira da Silva (isabelle.silva@unesp.br)
}
