% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{periodogram}
\alias{periodogram}
\title{Berechnung des Stichprobenspektrums}
\usage{
periodogram(ts, sr, type)
}
\arguments{
\item{ts}{}
}
\value{

}
\description{
periodogram berechnet das Stichprobenspektrum  (Schlittgen Seite 68) einer gegebenen Zeitreihe:
\deqn{I(\lambda) = N\cdot |D(\lambda)|^2}
}
