% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_estimators.R
\name{periodogram}
\alias{periodogram}
\title{Berechnung des Stichprobenspektrums}
\usage{
periodogram(data, sr, signal, type)
}
\arguments{
\item{sr}{sample-rate}

\item{ts}{datensatz mit mind den variablen t und signal}
}
\value{

}
\description{
periodogram berechnet das Stichprobenspektrum  (Schlittgen Seite 68) einer gegebenen Zeitreihe:
\deqn{I(\lambda) = N\cdot |D(\lambda)|^2}
\eqn{D(\lambda)}: Fouriertransformierte der Zeitreihe.
}
